Update key parsing and move fun
