+++ resources/certbuildutils.py
@@ -11,19 +11,15 @@ from typing import Any, List, Optional, Sequence, Set, Tuple, Union
 
 import pyasn1.error
 from cryptography import x509
-from cryptography.hazmat.primitives.serialization import Encoding, PublicFormat
 from pyasn1.codec.der import decoder, encoder
 from pyasn1.type import tag, univ, useful
 from pyasn1.type.base import Asn1Item, Asn1Type
-from pyasn1_alt_modules import rfc4211, rfc5280, rfc5480, rfc5652, rfc6402, rfc8954, rfc9480, rfc9481
+from pyasn1_alt_modules import rfc4211, rfc5280, rfc5652, rfc6402, rfc8954, rfc9480, rfc9481
 from pyasn1_alt_modules.rfc2459 import AttributeValue
 from robot.api.deco import keyword, not_keyword
 
-from pq_logic.keys.abstract_wrapper_keys import AbstractCompositePrivateKey
-from pq_logic.keys.composite_sig03 import CompositeSig03PrivateKey
-from pq_logic.keys.trad_kem_keys import RSAEncapKey
 from pq_logic.pq_utils import is_kem_public_key
-from pq_logic.tmp_oids import COMPOSITE_SIG03_HASH_OID_2_NAME, COMPOSITE_SIG04_HASH_OID_2_NAME, id_rsa_kem_spki
+from pq_logic.tmp_oids import COMPOSITE_SIG03_HASH_OID_2_NAME, COMPOSITE_SIG04_HASH_OID_2_NAME
 from resources import (
     asn1utils,
     certextractutils,
@@ -40,7 +36,6 @@ from resources import (
     utils,
 )
 from resources.asn1utils import get_set_bitstring_names
-from resources.convertutils import subject_public_key_info_from_pubkey
 from resources.copyasn1utils import copy_name
 from resources.exceptions import BadAsn1Data, BadCertTemplate
 from resources.oid_mapping import may_return_oid_to_name
@@ -1513,175 +1508,6 @@ def prepare_cert_template_from_csr(csr: rfc6402.CertificationRequest) -> rfc4211
     subject = utils.get_openssl_name_notation(csr["certificationRequestInfo"]["subject"])  # type: ignore
     subject: str
     return prepare_cert_template(key=public_key, subject=subject, extensions=extensions)
-
-
-@keyword(name="Prepare SubjectPublicKeyInfo")
-def prepare_subject_public_key_info(  # noqa D417 undocumented-param
-    key: Optional[Union[PrivateKey, PublicKey]] = None,
-    for_kga: bool = False,
-    key_name: Optional[str] = None,
-    use_rsa_pss: bool = False,
-    use_pre_hash: bool = False,
-    hash_alg: Optional[str] = None,
-    invalid_key_size: bool = False,
-    add_params_rand_bytes: bool = False,
-    add_null: bool = False,
-) -> rfc5280.SubjectPublicKeyInfo:
-    """Prepare a `SubjectPublicKeyInfo` structure for a `Certificate`, `CSR` or `CertTemplate`.
-
-    For invalid Composite keys must the private key be provided.
-
-    Note: If the key is a CompositeSig key, the `key_name` the private key must be provided,
-    if the RSA key has an invalid key size.
-
-    Arguments:
-    ---------
-        - `key`: The public or private key to use for the `SubjectPublicKeyInfo`.
-        - `for_kga`: A flag indicating whether the key is for a key generation authority (KGA).
-        - `key_name`: The key algorithm name to use for the `SubjectPublicKeyInfo`.
-        (can be set to `rsa_kem`. RFC9690). Defaults to `None`.
-        - `use_rsa_pss`: Whether to use RSA-PSS padding. Defaults to `False`.
-        - `use_pre_hash`: Whether to use the pre-hash version for a composite-sig key. Defaults to `False`.
-        - `hash_alg`: The pre-hash algorithm to use for the pq signature key. Defaults to `None`.
-        - `invalid_key_size`: A flag indicating whether the key size is invalid. Defaults to `False`.
-        - `add_params_rand_bytes`: A flag indicating whether to add random bytes to the key parameters. \
-        Defaults to `False`.
-        - `add_null`: A flag indicating whether to add a null value to the key parameters. Defaults to `False`.
-
-    Returns:
-    -------
-        - The populated `SubjectPublicKeyInfo` structure.
-
-    Raises:
-    ------
-        - `ValueError`: If no key is provided and the for_kga flag is not set.
-        - `ValueError`: If both `add_null` and `add_params_rand_bytes` are set.
-
-
-    Examples:
-    --------
-    | ${spki}= | Prepare SubjectPublicKeyInfo | key=${key} | use_rsa_pss=True |
-    | ${spki}= | Prepare SubjectPublicKeyInfo | key=${key} | key_name=rsa-kem |
-    | ${spki}= | Prepare SubjectPublicKeyInfo | key=${key} | for_kga=True |
-    | ${spki}= | Prepare SubjectPublicKeyInfo | key=${key} | add_null=True |
-
-    """
-    if key is None and not for_kga:
-        raise ValueError("Either a key has to be provided or the for_kga flag have to be set.")
-
-    if add_null and add_params_rand_bytes:
-        raise ValueError("Either `add_null` or `add_params_rand_bytes` can be set, not both.")
-
-    if isinstance(key, AbstractCompositePrivateKey):
-        pub_key = key.public_key().public_bytes(encoding=Encoding.DER, format=PublicFormat.Raw)
-        spki = rfc5280.SubjectPublicKeyInfo()
-        pub_key = pub_key if not invalid_key_size else pub_key + b"\x00"
-        spki["subjectPublicKey"] = univ.BitString.fromOctetString(pub_key)
-        if isinstance(key, CompositeSig03PrivateKey):
-            oid = key.get_oid(use_pss=use_rsa_pss, pre_hash=use_pre_hash)
-        else:
-            oid = key.get_oid()
-        spki["algorithm"]["algorithm"] = oid
-
-        if add_null:
-            spki["algorithm"]["parameters"] = univ.Null("")
-
-        if add_params_rand_bytes:
-            spki["algorithm"]["parameters"] = univ.BitString.fromOctetString(os.urandom(16))
-
-        return spki
-
-    if isinstance(key, PrivateKey):
-        key = key.public_key()
-
-    if for_kga:
-        return _prepare_spki_for_kga(
-            key=key,
-            key_name=key_name,
-            use_pss=use_rsa_pss,
-            use_pre_hash=use_pre_hash,
-            add_null=add_null,
-            add_params_rand_bytes=add_params_rand_bytes,
-        )
-
-    if key_name in ["rsa-kem", "rsa_kem"]:
-        key = RSAEncapKey(key)  # type: ignore
-
-    spki = subject_public_key_info_from_pubkey(
-        public_key=key,  # type: ignore
-        use_rsa_pss=use_rsa_pss,
-        use_pre_hash=use_pre_hash,
-        hash_alg=hash_alg,
-    )
-
-    if invalid_key_size:
-        tmp = spki["subjectPublicKey"].asOctets() + b"\x00\x00"
-        spki["subjectPublicKey"] = univ.BitString.fromOctetString(tmp)
-
-    if add_params_rand_bytes:
-        spki["algorithm"]["parameters"] = univ.BitString.fromOctetString(os.urandom(16))
-
-    return spki
-
-
-def _prepare_spki_for_kga(
-    key: Optional[Union[PrivateKey, PublicKey]] = None,
-    key_name: Optional[str] = None,
-    use_pss: bool = False,
-    use_pre_hash: bool = False,
-    add_null: bool = False,
-    add_params_rand_bytes: bool = False,
-) -> rfc5280.SubjectPublicKeyInfo:
-    """Prepare a SubjectPublicKeyInfo for KGA usage.
-
-    :param key: A private or public key.
-    :param key_name: An optional key algorithm name.
-    :param use_pss: Whether to use PSS padding for RSA and a RSA-CompositeKey.
-    :param use_pre_hash: Whether to use the pre-hash version for a composite-sig key. Defaults to `False`.
-    :param add_null: Whether to add a null value to the key parameters. Defaults to `False`.
-    :param add_params_rand_bytes: Whether to add random bytes to the key parameters. Defaults to `False`.
-    :return: The populated `SubjectPublicKeyInfo` structure.
-    """
-    if add_null and add_params_rand_bytes:
-        raise ValueError("Either `add_null` or `add_params_rand_bytes` can be set, not both.")
-
-    spki = rfc5280.SubjectPublicKeyInfo()
-    spki["subjectPublicKey"] = univ.BitString("")
-
-    if key is not None:
-        if isinstance(key, typingutils.PrivateKey):
-            key = key.public_key()
-
-    if key_name and key_name in ["rsa", "dsa", "ecc", "rsa-kem"]:
-        names_2_oid = {
-            "rsa": univ.ObjectIdentifier("1.2.840.113549.1.1.1"),
-            "dsa": univ.ObjectIdentifier("1.2.840.10040.4.1"),
-            "ecc": univ.ObjectIdentifier("1.2.840.10045.3.1.7"),
-            "rsa-kem": id_rsa_kem_spki,
-        }
-        spki["algorithm"]["algorithm"] = names_2_oid[key_name]
-        if key_name == "ecc":
-            spki["algorithm"]["parameters"] = rfc5480.ECParameters()
-            spki["algorithm"]["parameters"]["namedCurve"] = rfc5480.secp256r1
-
-    if key_name is not None:
-        from pq_logic.combined_factory import CombinedKeyFactory
-
-        key = CombinedKeyFactory.generate_key(key_name).public_key()
-        spki_tmp = subject_public_key_info_from_pubkey(public_key=key, use_rsa_pss=use_pss, use_pre_hash=use_pre_hash)
-        spki["algorithm"]["algorithm"] = spki_tmp["algorithm"]["algorithm"]
-
-    elif key is not None:
-        spki_tmp = subject_public_key_info_from_pubkey(public_key=key, use_rsa_pss=use_pss)
-        spki["algorithm"]["algorithm"] = spki_tmp["algorithm"]["algorithm"]
-
-    if add_null:
-        spki["algorithm"]["parameters"] = univ.Null("")
-
-    if add_params_rand_bytes:
-        spki["algorithm"]["parameters"] = univ.BitString.fromOctetString(os.urandom(16))
-
-    return spki
 
 
 @not_keyword
